Reorganize UI based on feedback from community
